+++ src/lib.rs
@@ -31,7 +31,10 @@
 
 #![no_std]
 
+mod kangaroo;
+
 use crunchy::unroll;
+pub use kangaroo::KangarooTwelve;
 
 const RHO: [u32; 24] = [
     1, 3, 6, 10, 15, 21, 28, 36, 45, 55, 2, 14, 27, 41, 56, 8, 25, 43, 62, 18, 39, 61, 20, 44,
@@ -68,30 +71,10 @@ const RC: [u64; 24] = [
     0x8000000080008008u64,
 ];
 
-const RC_KANGAROO: [u64; 12] = [
-	0x000000008000808b,
-	0x800000000000008b,
-	0x8000000000008089,
-	0x8000000000008003,
-	0x8000000000008002,
-	0x8000000000000080,
-	0x000000000000800a,
-	0x800000008000000a,
-	0x8000000080008081,
-	0x8000000000008080,
-	0x0000000080000001,
-	0x8000000080008008,
-];
-
 /// keccak-f[1600]
-#[inline]
-pub fn keccakf(a: &mut [u64; PLEN]) {
-    keccakf_with_rounds(a, &RC);
-}
-
 #[allow(unused_assignments)]
-fn keccakf_with_rounds(a: &mut [u64; PLEN], rounds: &[u64]) {
-    for round in rounds {
+pub fn keccakf(a: &mut [u64; PLEN]) {
+    for i in 0..24 {
         let mut array: [u64; 5] = [0; 5];
 
         // Theta
@@ -147,7 +130,7 @@ fn keccakf_with_rounds(a: &mut [u64; PLEN], rounds: &[u64]) {
         };
 
         // Iota
-        a[0] ^= round;
+        a[0] ^= RC[i];
     }
 }
 
@@ -212,6 +195,12 @@ impl Buffer {
     }
 }
 
+#[derive(Debug, Clone, Copy)]
+enum Rounds {
+    Normal,
+    Reduced,
+}
+
 /// This structure should be used to create keccak/sha3 hash.
 ///
 /// ```rust
@@ -245,7 +234,7 @@ pub struct Keccak {
     offset: usize,
     rate: usize,
     delim: u8,
-    rounds: &'static [u64],
+    rounds: Rounds,
 }
 
 impl Clone for Keccak {
@@ -294,10 +283,10 @@ impl_global_alias!(sha3_512, 512);
 
 impl Keccak {
     pub fn new(rate: usize, delim: u8) -> Keccak {
-        Keccak::new_with_rounds(rate, delim, &RC)
+        Keccak::new_with_rounds(rate, delim, Rounds::Normal)
     }
 
-    pub fn new_with_rounds(rate: usize, delim: u8, rounds: &'static [u64]) -> Keccak {
+    fn new_with_rounds(rate: usize, delim: u8, rounds: Rounds) -> Keccak {
         assert!(rate != 0, "rate cannot be equal 0");
         Keccak {
             buffer: Buffer::default(),
@@ -325,7 +314,10 @@ impl Keccak {
 
     #[inline]
     pub fn keccakf(&mut self) {
-        keccakf_with_rounds(self.buffer.inner(), self.rounds);
+        match self.rounds {
+            Rounds::Normal => keccakf(self.buffer.inner()),
+            Rounds::Reduced => kangaroo::keccakf(self.buffer.inner()),
+        }
     }
 
     pub fn finalize(mut self, output: &mut [u8]) {
@@ -422,99 +414,3 @@ impl XofReader {
     }
 }
 
-/// KangarooTwelve's length encoding.
-struct EncodedLen {
-    offset: usize,
-    buffer: [u8; 9],
-}
-
-impl EncodedLen {
-    fn new(len: usize) -> Self {
-        let len_view = len.to_be_bytes();
-        let offset = len_view.iter().position(|i| *i != 0).unwrap_or(8);
-        let mut buffer = [0u8; 9];
-        buffer[..8].copy_from_slice(&len_view);
-        buffer[8] = 8 - offset as u8;
-
-        EncodedLen {
-            offset,
-            buffer,
-        }
-    }
-
-    fn value(&self) -> &[u8] {
-        &self.buffer[self.offset..]
-    }
-}
-
-pub struct KangarooTwelve<T> {
-    state: Keccak,
-    current_chunk: Keccak,
-    custom_string: Option<T>,
-    written: usize,
-    chunks: usize,
-}
-
-impl<T: AsRef<[u8]>> KangarooTwelve<T> {
-    const MAX_CHUNK_SIZE: usize = 8192;
-
-    pub fn new(custom_string: T) -> Self {
-        KangarooTwelve {
-            state: Keccak::new_with_rounds(168, 0, &RC_KANGAROO),
-            current_chunk: Keccak::new_with_rounds(168, 0x0b, &RC_KANGAROO),
-            custom_string: Some(custom_string),
-            written: 0,
-            chunks: 0,
-        }
-    }
-
-    pub fn update(&mut self, input: &[u8]) {
-        let mut to_absorb = input;
-        while to_absorb.len() > 0 {
-            if self.written == Self::MAX_CHUNK_SIZE {
-                if self.chunks == 0 {
-                    self.state.update(&[0x03, 0, 0, 0, 0, 0, 0, 0]);
-                } else {
-                    let mut tmp_chunk = [0u8; 32];
-                    self.current_chunk.clone().finalize(&mut tmp_chunk);
-                    self.state.update(&tmp_chunk);
-                    self.current_chunk = Keccak::new_with_rounds(168, 0x0b, &RC_KANGAROO);
-                }
-
-                self.written = 0;
-                self.chunks += 1;
-            }
-
-            let todo = ::core::cmp::min(Self::MAX_CHUNK_SIZE - self.written, to_absorb.len());
-            if self.chunks == 0 {
-                self.state.update(&to_absorb[..todo]);
-            } else {
-                self.current_chunk.update(&to_absorb[..todo]);
-            }
-            self.written += todo;
-            to_absorb = &to_absorb[todo..];
-        }
-    }
-
-    pub fn finalize(mut self, output: &mut [u8]) {
-        let custom_string = self.custom_string.take()
-            .expect("KangarooTwelve cannot be initialized without custom_string; qed");
-        let encoded_len = EncodedLen::new(custom_string.as_ref().len());
-        self.update(custom_string.as_ref());
-        self.update(encoded_len.value());
-
-        if self.chunks == 0 {
-            self.state.delim = 0x07;
-        } else {
-            let encoded_chunks = EncodedLen::new(self.chunks);
-            let mut tmp_chunk = [0u8; 32];
-            self.current_chunk.finalize(&mut tmp_chunk);
-            self.state.update(&tmp_chunk);
-            self.state.update(encoded_chunks.value());
-            self.state.update(&[0xff, 0xff]);
-            self.state.delim = 0x06;
-        }
-
-        self.state.finalize(output);
-    }
-}
